+++ dashboard.html
@@ -14,7 +14,6 @@
         body {
             background-color: #f8f9fa;
             transition: all 0.5s ease;
-            padding-bottom: 2rem;
         }
 
         /* Контент сжимаем только при ширине окна больше 1000px */
@@ -23,7 +22,6 @@
             max-width: 80%;
             margin: 0 auto;
             transition: all 0.5s ease;
-            padding-bottom: 2rem;
         }
 
         /* Для узких окон (<1000px) делаем контент "жидким" */
@@ -98,10 +96,6 @@
                 <h3 class="card-title text-danger mb-2">
                     Phishing Awareness Dashboard
                 </h3>
-                <div class="d-flex gap-2 mb-2">
-                    <a href="/" class="btn btn-danger btn-sm">Dashboard</a>
-                    <a href="/emails" class="btn btn-outline-primary btn-sm">Email адреса</a>
-                </div>
                 <p class="mb-1"><b>Пример ссылки:</b></p>
                 <code>http://SERVER_IP:8080/track/user123</code>
             </div>
